# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mass Finder API",
        "description": "API para busca de paróquias e horários de missa",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "{baseUrl}/api",
            "description": "API Base URL",
            "variables": {
                "baseUrl": {
                    "default": "http://localhost:8000"
                }
            }
        }
    ],
    "paths": {
        "/admin/parishes/{parish}/mass-schedules": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Listar horários de missa da paróquia",
                "operationId": "9078dbf835cc78dde0dfd6b4101a9ec7",
                "parameters": [
                    {
                        "name": "parish",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de horários",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/MassSchedule"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenericMessageResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Sem permissão",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenericMessageResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Paróquia não encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenericMessageResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Criar horário de missa para paróquia",
                "operationId": "11f47fa1064d781716b6b5e252533937",
                "parameters": [
                    {
                        "name": "parish",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateMassScheduleRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Horário criado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/MassSchedule"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenericMessageResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Sem permissão",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenericMessageResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Paróquia não encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenericMessageResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validação falhou",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorValidationResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/mass-schedules/{massSchedule}": {
            "put": {
                "tags": [
                    "Admin"
                ],
                "summary": "Atualizar horário de missa",
                "operationId": "c5cb513f877944265307a0d79a5f5c50",
                "parameters": [
                    {
                        "name": "massSchedule",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateMassScheduleRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Horário atualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/MassSchedule"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenericMessageResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Sem permissão",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenericMessageResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Horário não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenericMessageResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validação falhou",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorValidationResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Remover horário de missa",
                "operationId": "94c93f426bc3d3200dd283fd8b54427b",
                "parameters": [
                    {
                        "name": "massSchedule",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Horário removido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenericMessageResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenericMessageResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Sem permissão",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenericMessageResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Horário não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenericMessageResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "MassSchedule": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "parish_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "day_of_week": {
                        "description": "0=Domingo, 1=Segunda, ..., 6=Sábado",
                        "type": "integer",
                        "maximum": 6,
                        "minimum": 0
                    },
                    "time": {
                        "type": "string",
                        "example": "08:00"
                    },
                    "notes": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "CreateMassScheduleRequest": {
                "required": [
                    "day_of_week",
                    "time"
                ],
                "properties": {
                    "day_of_week": {
                        "description": "0=Domingo, 1=Segunda, ..., 6=Sábado",
                        "type": "integer",
                        "maximum": 6,
                        "minimum": 0
                    },
                    "time": {
                        "type": "string",
                        "pattern": "^([01]?[0-9]|2[0-3]):[0-5][0-9]$",
                        "example": "08:00"
                    },
                    "notes": {
                        "type": "string",
                        "maxLength": 500,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "GenericMessageResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Operação realizada com sucesso."
                    }
                },
                "type": "object"
            },
            "ErrorValidationResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token obtido via POST /api/auth/login",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}